# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Abrpardaz API",
        "description": "This is the official API which can be used by any user",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api/v1",
            "description": "Local server"
        },
        {
            "url": "http://abrpardaz.com/api/v1",
            "description": "The API endpoint"
        }
    ],
    "paths": {
...
        "/machines/createFromImage": {
            "post": {
                "tags": [
                    "Machine"
                ],
                "summary": "Create a new machine from an image",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\MachineController::createFromImage",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/machines/{id}/console": {
            "get": {
                "tags": [
                    "Machine"
                ],
                "summary": "Get console url of the machine ",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\MachineController::console",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/machines/{id}/powerOn": {
            "post": {
                "tags": [
                    "Machine"
                ],
                "summary": "powers on the machine ",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\MachineController::powerOn",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/machines/{id}/powerOff": {
            "post": {
                "tags": [
                    "Machine"
                ],
                "summary": "powers off the machine ",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\MachineController::powerOff",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/machines/{id}/takeSnapshot": {
            "post": {
                "tags": [
                    "Machine"
                ],
                "summary": "Take snapshot from a machine",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\MachineController::takeSnapshot",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name you want to put on the snapshot",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/machines/{id}/resendInfo": {
            "put": {
                "tags": [
                    "Machine"
                ],
                "summary": "Resends information of this machine",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\MachineController::resendInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/machines/{id}/rename": {
            "post": {
                "tags": [
                    "Machine"
                ],
                "summary": "Change the name of a machine",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\MachineController::rename",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The new name you want to put on the machine",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
        "/machines/{id}/remove": {
            "delete": {
                "tags": [
                    "Machine"
                ],
                "summary": "Removes the machine ",
                "description": "",
                "operationId": "App\\Http\\Controllers\\V1\\MachineController::remove",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "",
                        "required": true,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": ""
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Login and copy the token into the below box",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}